# internal/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: SWAMP API
  description: Software Assurance Marketplace — AI-powered security analysis of Git repositories.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current session and user
      operationId: getCurrentSession
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/auth/oidc/callback:
    get:
      tags: [auth]
      summary: OIDC callback
      operationId: oidcCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to application
  /api/v1/auth/agree-aup:
    post:
      tags: [auth]
      summary: Accept the Acceptable Use Policy
      operationId: agreeAUP
      security:
        - cookieAuth: []
      responses:
        "200":
          description: AUP accepted

  # ── Groups ────────────────────────────────────────────────
...
  /api/v1/admin/users/{userID}:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [admin]
      summary: Get a user
      operationId: getUser
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      tags: [admin]
      summary: Update a user
      operationId: updateUser
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                email:
                  type: string
                status:
                  type: string
                  enum: [active, suspended, disabled]
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags: [admin]
      summary: Delete a user
      operationId: deleteUser
      security:
        - cookieAuth: []
      responses:
        "204":
          description: User deleted
  /api/v1/admin/users/{userID}/roles:
    parameters:
      - $ref: "#/components/parameters/userID"
    post:
      tags: [admin]
      summary: Add a role to a user
      operationId: addUserRole
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
      responses:
        "201":
          description: Role added
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: swamp_session
    bearerAuth:
      type: http
      scheme: bearer
      description: API key in Authorization header
  parameters:
...
    userID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    SessionInfo:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        roles:
          type: array
          items:
            type: string
        aup_agreed:
          type: boolean
        expires_at:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, suspended, disabled]
        last_login:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...